added masked decompression operators
